refactor: 인터셉터 리팩토링
