+++ components/module-oauth/src/test/java/depromeet/ohgzoo/iam/oauth/OauthServiceImplTest.java
@@ -1,6 +1,7 @@
 package depromeet.ohgzoo.iam.oauth;
 
 import depromeet.ohgzoo.iam.jwt.SpyJwtService;
+import depromeet.ohgzoo.iam.jwt.UnAuthenticationException;
 import depromeet.ohgzoo.iam.oauth.kakao.KakaoProfileResponse;
 import depromeet.ohgzoo.iam.oauth.kakao.KakaoProfileResponse.KakaoAccount;
 import depromeet.ohgzoo.iam.oauth.kakao.KakaoTokenResponse;
