+++ components/support-jwt/src/main/java/depromeet/ohgzoo/iam/jwt/UnAuthenticationException.java
@@ -1,4 +1,4 @@
-package depromeet.ohgzoo.iam.oauth;
+package depromeet.ohgzoo.iam.jwt;
 
 public class UnAuthenticationException extends RuntimeException {
     public UnAuthenticationException(String message) {
